requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-allauth[socialaccount]>=0.57.0
django-cors-headers>=4.3.0
python-decouple>=3.8
qrcode[pil]>=7.4.2
Pillow>=10.0.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
